separate daily puzzle button code
