+++ src/components/home/Home.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useRef, useState, useMemo } from "react";
+import { useMemo } from "react";
 import { Link } from "react-router-dom";
 import {
   FaDiscord as DiscordIcon,
@@ -8,99 +8,7 @@ import { GamemodeButtonProps } from "./GamemodeButton";
 import { GamemodeSection } from "./GamemodeSection";
 import "./Home.css";
 
-const second = 1000;
-const minute = second * 60;
-const hour = minute * 60;
-const day = hour * 24;
-
-const formatCountdownDistance = (distance: number): string => {
-  const hours = Math.floor((distance % day) / hour)
-    .toString()
-    .padStart(2, "0");
-  const minutes = Math.floor((distance % hour) / minute)
-    .toString()
-    .padStart(2, "0");
-  const seconds = Math.floor((distance % minute) / second)
-    .toString()
-    .padStart(2, "0");
-
-  return `${hours}:${minutes}:${seconds}`;
-};
-
-function convertDateToUTC(date: Date) {
-  return new Date(
-    date.getUTCFullYear(),
-    date.getUTCMonth(),
-    date.getUTCDate(),
-    date.getUTCHours(),
-    date.getUTCMinutes(),
-    date.getUTCSeconds()
-  );
-}
-
 export const Home = (): JSX.Element => {
-  const [dailyPuzzleSeed, setDailyPuzzleSeed] = useState<string | undefined>();
-  const [now, setNow] = useState<Date | null>(null);
-  const [end, setEnd] = useState<Date | null>(null);
-  const intervalRef = useRef<NodeJS.Timer | null>(null);
-
-  async function updateDailyPuzzle() {
-    const resp = await fetch("../../api/daily.php");
-    if (resp.ok) {
-      const text = await resp.text();
-      if (text.length > 20) {
-        // Somehow launching this in local environment returns index.html?
-        console.error("Failed to fetch daily puzzle");
-      } else {
-        setDailyPuzzleSeed(text);
-      }
-    } else {
-      console.error("Failed to fetch daily puzzle");
-    }
-  }
-
-  function updateCountdownEnd() {
-    const nowUTC = convertDateToUTC(new Date());
-    const endUTC = new Date(nowUTC);
-    endUTC.setDate(endUTC.getDate() + 1);
-    endUTC.setHours(0, 0, 0, 0);
-    setEnd(endUTC);
-  }
-
-  useEffect(() => {
-    updateDailyPuzzle();
-  }, []);
-
-  useEffect(() => {
-    setNow(convertDateToUTC(new Date()));
-    updateCountdownEnd();
-
-    intervalRef.current = setInterval(() => {
-      setNow(convertDateToUTC(new Date()));
-    }, 1000);
-
-    return () => {
-      if (intervalRef.current) {
-        clearInterval(intervalRef.current);
-      }
-    };
-  }, []);
-
-  useEffect(() => {
-    if (now !== null && end !== null) {
-      if (now > end) {
-        updateCountdownEnd();
-        updateDailyPuzzle();
-      }
-    }
-  }, [now, end]);
-
-  let dailyTimeRemaining = "";
-  if (now !== null && end !== null) {
-    const distance = end.getTime() - now.getTime();
-    dailyTimeRemaining = formatCountdownDistance(distance);
-  }
-
   const puzzleBoards: GamemodeButtonProps[] = useMemo(
     () => [
       {
@@ -225,24 +133,8 @@ export const Home = (): JSX.Element => {
         title="Puzzle"
         description="You must find where the pieces are without being allowed to dig"
         buttons={puzzleBoards}
+        includeDaily
       />
-      <div className="flex hor">
-        {dailyPuzzleSeed && (
-          <Link
-            className="gamemode button"
-            to={`/play?g=p&p=R3B3N3K1&s=10&c=8&r=${dailyPuzzleSeed}`}
-            id="daily"
-          >
-            <h2>Daily {dailyTimeRemaining}</h2>
-            8 pieces
-            <br />
-            Rook, Bishop, Knight and King
-            <br />
-            10x10
-          </Link>
-        )}
-      </div>
-
       <GamemodeSection
         title="Classic"
         description="Dig and try to find where the pieces are"
